Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: 1aae36f4813064f225c5a6fff4998ed5455515c5 [formerly 26ade7cae88d84ef9e1aed95b3df2cd6d5f893ee] [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6] [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6 [formerly aa991c922354f0a67fac08fee456ca60a30de968]] [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6 [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]]]]
Former-commit-id: 74a6c3dd20898e9bd4484376ea4e4e471bdec9af
Former-commit-id: 0a22b6429b06d5e570b21625a0de6b06f911d381 [formerly 53a85ec6b9da7b5e37777469d995ba034e153ce6] [formerly 5977b049c1eda298ce206a8d5fc34b8f235639c5 [formerly 10ea02b9366284a7b97f57ffdb7ac54688cfa0c4]]
Former-commit-id: 5977b049c1eda298ce206a8d5fc34b8f235639c5
Former-commit-id: d395fc27fc163b2f1f124017e47c57f63e8859d1

+++ edexOsgi/com.raytheon.uf.edex.plugin.vaa/src/com/raytheon/uf/edex/plugin/vaa/decoder/VAAParser.java
@@ -51,6 +51,7 @@ import com.raytheon.uf.common.wmo.WMOHeader;
  * Feb 11, 2014 2763       skorolev    Made LFCR correction of input data.
  * Mar 10, 2014 2807       skorolev    Added MalformedDataException for VAA decoding.
  * May 14, 2014 2536       bclement    moved WMO Header to common
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -59,11 +60,10 @@ import com.raytheon.uf.common.wmo.WMOHeader;
  */
 public class VAAParser implements Iterable<VAARecord> {
 
-    // TODO should use JTS coordinate
     private static class LatLon {
-        public Double lat;
+        public float lat;
 
-        public Double lon;
+        public float lon;
 
         @Override
         public String toString() {
@@ -266,12 +266,12 @@ public class VAAParser implements Iterable<VAARecord> {
         Matcher m = InternalReport.LAT_LON_P.matcher(latLon);
         if (m.find()) {
             latlon = new LatLon();
-            latlon.lat = Double.parseDouble(m.group(2));
-            latlon.lat += (Double.parseDouble(m.group(3)) / 60.0);
+            latlon.lat = Float.parseFloat(m.group(2));
+            latlon.lat += (Float.parseFloat(m.group(3)) / 60.0f);
             latlon.lat *= ("S".equals(m.group(1))) ? -1 : 1;
 
-            latlon.lon = Double.parseDouble(m.group(5));
-            latlon.lon += (Double.parseDouble(m.group(6)) / 60.0);
+            latlon.lon = Float.parseFloat(m.group(5));
+            latlon.lon += (Float.parseFloat(m.group(6)) / 60.0f);
             latlon.lon *= ("W".equals(m.group(4))) ? -1 : 1;
         }
         return latlon;
